pyproject.toml
[tool.poetry]
name = "af-tools"
version = "0.1.0"
description = ""
authors = ["ahmedselim2017 <ahmedselimuzum@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"

matplotlib = "^3.9.0"
orjson = "^3.10.3"
biopython = "^1.83"
scikit-learn = "^1.5.0"
tqdm = "^4.66.4"
[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
af_tools = "af_tools.cli:cli"
